fix startup crash

+++ src/main.cpp
@@ -4,7 +4,9 @@
 using namespace geode::prelude;
 
 $on_mod(Loaded) {
-    for (auto file : std::filesystem::directory_iterator(dirs::getTempDir() / "profiles")) {
-        std::filesystem::remove_all(file.path());
+    if (std::filesystem::exists(dirs::getTempDir() / "profiles")) {
+        for (auto file : std::filesystem::directory_iterator(dirs::getTempDir() / "profiles")) {
+            std::filesystem::remove_all(file.path());
+        }
     }
 }
